+++ docs/api/str/len.rst
@@ -1,6 +1,7 @@
 
 .. xfunction:: datatable.str.len
     :src: src/core/expr/fexpr_slice.cc FExpr_Slice::evaluate_n
+    :tests: tests/expr/str/test-len.py
     :cvar: doc_str_len
     :signature: len(column)
