+++ tests/ijby/test-func-unary.py
@@ -220,35 +220,3 @@ def test_log_srcs(src, fn):
              mathlog(x)
              for x in src]
     assert DT1.to_list()[0] == pyans
-
-
-
-#-------------------------------------------------------------------------------
-# len()
-#-------------------------------------------------------------------------------
-
-def test_len():
-    DT = dt.Frame(A=["", "one", "2", "three", "four", None, "six", "seventy"])
-    RES = DT[:, f.A.len()]
-    assert RES.stype == dt.stype.int64
-    assert RES.to_list() == [[0, 3, 1, 5, 4, None, 3, 7]]
-
-
-def test_len2():
-    DT = dt.Frame([None, "", "mooo" * 10000], stype="str64")
-    RES = DT[:, f[0].len()]
-    assert RES.stype == dt.stype.int64
-    assert RES.to_list() == [[None, 0, 40000]]
-
-
-def test_len_wrong_col():
-    DT = dt.Frame(range(34))
-    with pytest.raises(TypeError, match="Function len cannot be applied to "
-                                        "a column of type int32"):
-        assert DT[:, f[0].len()]
-
-
-def test_len_unicode():
-    DT = dt.Frame(["майдан", "蒙蒂巨蟒", "🤥", "𝔘𝔫𝔦𝔠𝔬𝔡𝔢"])
-    RES = DT[:, f[0].len()]
-    assert_equals(RES, dt.Frame([6, 4, 1, 7], stype=dt.int64))
